+++ app/src/main/java/com/example/movedistance/BTSProcessor.java
@@ -69,9 +69,9 @@ public class BTSProcessor {
             before.set(uniq);
         }
 
-        int total = uniqs.stream().flatMap(Set::stream).collect(Collectors.toSet()).size();
-        int jerk_min = jerkList.isEmpty() ? -1 : Collections.min(jerkList);
-        int jerk_max = jerkList.isEmpty() ? -1 : Collections.max(jerkList);
+        float total = uniqs.stream().flatMap(Set::stream).collect(Collectors.toSet()).size();
+        float jerk_min = jerkList.isEmpty() ? -1 : Collections.min(jerkList);
+        float jerk_max = jerkList.isEmpty() ? -1 : Collections.max(jerkList);
         double jerk_mean = jerkList.isEmpty() ? -1.0 : jerkList.stream().mapToInt(Integer::intValue).average().orElse(-1.0);
         double jerk_std = jerkList.isEmpty() ? -1.0 : Math.sqrt(jerkList.stream().mapToDouble(j -> Math.pow(j - jerk_mean, 2)).average().orElse(-1.0));
